Update libxml.sh
update to be more plowly

+++ libxml.sh
@@ -1,1 +1,5 @@
-apt-get install libxslt-dev libxml2-dev -y
+if dpkg -s "libxml2-dev"; then
+  echo 'libxml2 already installed'
+else
+  apt-get install libxslt-dev libxml2-dev -y
+fi
